fix: allow parsing with straggling formatting
Prior to this, a comment line after the molecular formula is not
captured in the cleaning functions, and ends up contaminating the rest
of the process.

The fix here was to add an additional case in the final step looking for
opening HTML tags (i.e. "<"), and just doing a split to remove
everything that comes after it.